Major refactor to modularise graph running
This now has 3 functions, one to run while we're in the filter stage, one to run while we're in the main flowchart stage, and one that checks terminal node conditions and makes appropriate annotations based on that.

Thisis working in simple local tests, but has not been tested in anger yet. There's a risk the annotations will not come out properly, but we will have to see.